+++ KeyboardKitDemoKeyboard_SwiftUI/KeyboardViewController.swift
@@ -45,6 +45,7 @@ class KeyboardViewController: KeyboardInputViewController {
             changeKeyboardTypeAction: changeKeyboardType,
             toastContext: toastContext)
         keyboardLayoutProvider = StandardKeyboardLayoutProvider(
+            context: keyboardContext,
             inputSetProvider: keyboardInputSetProvider,
             leftSpaceAction: .keyboardType(.emojis),
             rightSpaceAction: .keyboardType(.images))